backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic[email]==2.9.0
pydantic-settings==2.5.0
redis==5.1.0
litellm==1.48.0
PyMuPDF==1.24.0
pdfplumber==0.11.0
python-docx==1.1.0
python-pptx==1.0.0
Pillow==10.4.0
pytesseract==0.3.13
boto3==1.35.0
apscheduler==3.10.4
slowapi==0.1.9
cryptography==43.0.0
websockets==13.0
httpx==0.27.0
